requirements.txt
streamlit==1.28.1
openai-whisper==20231117
google-generativeai==0.3.0
python-dotenv==1.0.0
